Move schemaStorageAccountCorsRule to helpers package
We need to call schemaStorageAccountCorsRule from package other than
storage, so we're moving it on its own package that can be imported from
two places.
